backend/requirements.txt
# Core FastAPI dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=1.10.7
pydantic-settings>=0.2.5

# Database and ORM
sqlalchemy[asyncio]>=2.0.20
alembic>=1.11.3
asyncpg>=0.28.0
psycopg2-binary>=2.9.7

# Redis and Celery
redis>=5.0.0
celery[redis]>=5.3.1

# S3 storage
boto3>=1.28.0
minio

# Authentication
python-jose[cryptography]>=3.3.0
cryptography>=41.0.3

# File handling
python-multipart>=0.0.6
pillow>=10.0.0

# Geospatial
geoalchemy2>=0.13.0
shapely>=2.0.1

# Machine learning
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0

# OCR and computer vision
pytesseract>=0.3.10
opencv-python-headless>=4.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.9.0
isort>=5.12.0

# Environment
python-dotenv>=1.0.0